+++ src/main/java/com/almostreliable/merequester/Registration.java
@@ -2,6 +2,7 @@ package com.almostreliable.merequester;
 
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceKey;
+import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.Item;
@@ -9,6 +10,8 @@ import net.minecraft.world.level.block.entity.BlockEntityType;
 
 import com.almostreliable.merequester.requester.RequesterBlock;
 import com.almostreliable.merequester.requester.RequesterBlockEntity;
+import com.almostreliable.merequester.requester.RequesterMenu;
+import com.almostreliable.merequester.terminal.RequesterTerminalMenu;
 import com.almostreliable.merequester.terminal.RequesterTerminalPart;
 
 import appeng.api.AECapabilities;
@@ -17,6 +20,7 @@ import appeng.block.AEBaseBlock;
 import appeng.blockentity.AEBaseBlockEntity;
 import appeng.items.parts.PartItem;
 import appeng.items.parts.PartModelsHelper;
+import net.neoforged.bus.api.IEventBus;
 import net.neoforged.neoforge.capabilities.RegisterCapabilitiesEvent;
 import net.neoforged.neoforge.event.BuildCreativeModeTabContentsEvent;
 import net.neoforged.neoforge.registries.DeferredBlock;
@@ -30,10 +34,14 @@ import static com.almostreliable.merequester.MERequester.TERMINAL_ID;
 
 public final class Registration {
 
-    public static final DeferredRegister.Blocks BLOCKS = DeferredRegister.createBlocks(BuildConfig.MOD_ID);
-    public static final DeferredRegister.Items ITEMS = DeferredRegister.createItems(BuildConfig.MOD_ID);
-    public static final DeferredRegister<BlockEntityType<?>> BLOCK_ENTITY_TYPES = DeferredRegister.create(
+    private static final DeferredRegister.Blocks BLOCKS = DeferredRegister.createBlocks(BuildConfig.MOD_ID);
+    private static final DeferredRegister.Items ITEMS = DeferredRegister.createItems(BuildConfig.MOD_ID);
+    private static final DeferredRegister<BlockEntityType<?>> BLOCK_ENTITIES = DeferredRegister.create(
         Registries.BLOCK_ENTITY_TYPE,
+        BuildConfig.MOD_ID
+    );
+    private static final DeferredRegister<MenuType<?>> MENUS = DeferredRegister.create(
+        Registries.MENU,
         BuildConfig.MOD_ID
     );
 
@@ -42,11 +50,26 @@ public final class Registration {
         RequesterBlock::new,
         AEBaseBlock.metalProps()
     );
-    public static final DeferredItem<BlockItem> REQUESTER_BLOCK_ITEM = ITEMS.registerSimpleBlockItem(
+    public static final DeferredItem<BlockItem> REQUESTER_ITEM = ITEMS.registerSimpleBlockItem(
         REQUESTER_ID,
         REQUESTER_BLOCK,
         new Item.Properties()
     );
+    public static final DeferredHolder<BlockEntityType<?>, BlockEntityType<RequesterBlockEntity>> REQUESTER_ENTITY
+        = BLOCK_ENTITIES.register(REQUESTER_ID, () -> {
+        // noinspection DataFlowIssue
+        var type = BlockEntityType.Builder
+            .of(RequesterBlockEntity::new, REQUESTER_BLOCK.get())
+            .build(null);
+        AEBaseBlockEntity.registerBlockEntityItem(type, REQUESTER_BLOCK.asItem());
+        REQUESTER_BLOCK.get().setBlockEntity(RequesterBlockEntity.class, type, null, null);
+        return type;
+    });
+    public static final DeferredHolder<MenuType<?>, MenuType<RequesterMenu>> REQUESTER_MENU = MENUS.register(
+        REQUESTER_ID,
+        () -> RequesterMenu.TYPE
+    );
+
     public static final DeferredItem<PartItem<RequesterTerminalPart>> REQUESTER_TERMINAL = ITEMS.registerItem(
         TERMINAL_ID,
         properties -> {
@@ -59,16 +82,17 @@ public final class Registration {
             );
         }
     );
+    public static final DeferredHolder<MenuType<?>, MenuType<RequesterTerminalMenu>> REQUESTER_TERMINAL_MENU = MENUS.register(
+        TERMINAL_ID,
+        () -> RequesterTerminalMenu.TYPE
+    );
 
-    public static final DeferredHolder<BlockEntityType<?>, BlockEntityType<RequesterBlockEntity>> REQUESTER_ENTITY
-        = BLOCK_ENTITY_TYPES.register(REQUESTER_ID, () -> {
-        var type = BlockEntityType.Builder
-            .of(RequesterBlockEntity::new, REQUESTER_BLOCK.get())
-            .build(null);
-        AEBaseBlockEntity.registerBlockEntityItem(type, REQUESTER_BLOCK.asItem());
-        REQUESTER_BLOCK.get().setBlockEntity(RequesterBlockEntity.class, type, null, null);
-        return type;
-    });
+    static void init(IEventBus modEventBus) {
+        BLOCKS.register(modEventBus);
+        ITEMS.register(modEventBus);
+        BLOCK_ENTITIES.register(modEventBus);
+        MENUS.register(modEventBus);
+    }
 
     private Registration() {}